Change migrations to store qty as a real number

+++ migrations/004.do.create-recipe-ingredients-table.sql
@@ -3,6 +3,6 @@ DROP TABLE IF EXISTS recipe_ingredients;
 CREATE TABLE recipe_ingredients (
     recipe_id INTEGER REFERENCES recipes(id),
     item_id INTEGER REFERENCES items(id),
-    qty INTEGER NOT NULL,
+    qty REAL NOT NULL,
     unit unit NOT NULL
 )
