+++ migrations/006.do.create-shopping-list-table.sql
@@ -2,6 +2,6 @@ DROP TABLE IF EXISTS shopping_list;
 
 CREATE TABLE shopping_list (
     item_id INTEGER REFERENCES items(id),
-    qty INTEGER NOT NULL,
+    qty REAL NOT NULL,
     unit unit NOT NULL  
 )
